chore(engine): extract dispatch switch from server.go
server.go had grown to exactly 800 lines (the Go hard cap), leaving
zero headroom: the next wire command added to the dispatch switch would
trip the file-size CI gate. The dispatch() switch is the highest-churn
surface in the package — every command adds a case.

Move dispatch() verbatim into a new dispatch.go sibling in the same
package (comments preserved, no behavior change). server.go drops from
800 to 420 lines; dispatch.go is 408. Prune the now-unused conversation
and providers imports from server.go.

Verified: go build, gofmt, golangci-lint, go test ./internal/server/,
and make check-file-sizes all pass.
